🔥 Remove Form.js and fixed typos
