+++ src/App.scss
@@ -1,4 +1,4 @@
-.Form {
+.App {
   width: 60rem;
   .heading {
     // text-align: center;
@@ -21,6 +21,30 @@
     margin: 0 auto;
   }
 }
+
+.CvPage {
+  border: 1px solid gray;
+  width: 70rem;
+  margin-top: 1rem;
+  @media (min-width: 80rem) {
+    position: absolute;
+    right: 2rem;
+    top: 2rem;
+  }
+  .cv-header {
+    background-color: black;
+    color: white;
+    padding: 2.5rem;
+    div {
+      margin: 0.5rem;
+    }
+    // .name {
+    //   font-size: 2.5rem;
+    //   font-weight: bold;
+    // }
+  }
+}
+
 // .Form {
 //   width: 30rem;
 //   background-color: #f8f5f1;
@@ -59,27 +83,5 @@
 //   }
 //   .heading {
 //     margin: 1rem 0;
-//   }
-// }
-
-// .CvPage {
-//   border: 1px solid red;
-//   width: 40rem;
-//   margin-top: 1rem;
-//   // position: absolute;
-//   right: 2rem;
-//   top: 2rem;
-//   height: 40rem;
-//   .cv-header {
-//     background-color: black;
-//     color: white;
-//     padding: 2.5rem;
-//     div {
-//       margin: 0.5rem;
-//     }
-//     // .name {
-//     //   font-size: 2.5rem;
-//     //   font-weight: bold;
-//     // }
 //   }
 // }
